Updated the code to fully support ptrs
The constructor in this branch is designed to take a vector of pointers.
This introduces some extra work because for manipulation of the data
dereferencing has to happen element wise for each vector. The ptrception
branch reduces this work by accepting a pointer to a vector. (Though it
just moves the work to someone else)